Refactor and clean up progress code
